[SYNTH-12989] Unexport `waitForResults()`
